resume_screening/requirements.txt
streamlit
pypdf
python-dotenv
google-generativeai
langchain
langchain-google-genai
langchain-community
chromadb
